+++ lib/rea.dart
@@ -1,0 +1,2 @@
+export 'src/git_service.dart';
+export 'src/analyzer.dart';