+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
-
-- Initial version.
+## 0.0.2
+- Added Phase 1: Core Analysis
+    - by default analyzes last 50 commits
+    - shows top 5 most changed files
+    - shows total files changed
+    - shows entropy
+    
+    - can pass number of last commits to analyze using --last option